Enforce Yarn usage and set engine versions
The changes primarily revolve around enforcing the usage of Yarn over npm and specifying the required versions of both. A `.npmrc` file has been added to strictly enforce engine versions, and `package.json` has been updated to specify that npm should not be used and to set Yarn's minimum version to 1.19.1.